New translation unit: Fanuc variables handler.
In the aim of decreasing of complexity.

	новый файл:    src/fanuc_vars.cpp
	изменено:      Makefile
	изменено:      src/ThreeWordFunc.cpp
	изменено:      src/translate_epp.cpp
	изменено:      src/translator.cpp
	изменено:      src/translator.h
